+++ Sources/ChessKit/Parsers/EngineLANParser.swift
@@ -17,9 +17,7 @@
 ///
 /// See [UCI protocol documentation](https://backscattering.de/chess/uci/2006-04.txt)
 /// for more information.
-public class EngineLANParser {
-
-    private init() {}
+public enum EngineLANParser {
 
     /// Parses a LAN string and returns a move.
     ///